+++ api/src/pdc/services/certificate/actions/CreateCertificateAction.integration.spec.ts
@@ -2,11 +2,14 @@
 import { faker } from "@/deps.ts";
 import { afterEach, assertEquals, beforeEach, describe, it, sinon } from "@/dev_deps.ts";
 import { ConfigInterfaceResolver, ContextType, KernelInterfaceResolver } from "@/ilos/common/index.ts";
-import { CarpoolInterface, CarpoolTypeEnum } from "@/shared/certificate/common/interfaces/CarpoolInterface.ts";
-import { CertificateBaseInterface } from "@/shared/certificate/common/interfaces/CertificateBaseInterface.ts";
-import { CertificateInterface } from "@/shared/certificate/common/interfaces/CertificateInterface.ts";
-import { CertificateMetaInterface } from "@/shared/certificate/common/interfaces/CertificateMetaInterface.ts";
-import { ParamsInterface, ResultInterface } from "@/shared/certificate/create.contract.ts";
+import {
+  CarpoolInterface,
+  CarpoolTypeEnum,
+} from "@/pdc/services/certificate/contracts/common/interfaces/CarpoolInterface.ts";
+import { CertificateBaseInterface } from "@/pdc/services/certificate/contracts/common/interfaces/CertificateBaseInterface.ts";
+import { CertificateInterface } from "@/pdc/services/certificate/contracts/common/interfaces/CertificateInterface.ts";
+import { CertificateMetaInterface } from "@/pdc/services/certificate/contracts/common/interfaces/CertificateMetaInterface.ts";
+import { ParamsInterface, ResultInterface } from "@/pdc/services/certificate/contracts/create.contract.ts";
 import { mapFromCarpools } from "../helpers/mapFromCarpools.ts";
 import { CarpoolRepositoryProviderInterfaceResolver } from "../interfaces/CarpoolRepositoryProviderInterface.ts";
 import { CertificateRepositoryProviderInterfaceResolver } from "../interfaces/CertificateRepositoryProviderInterface.ts";
