+++ api/src/pdc/services/certificate/actions/FindCertificateAction.ts
@@ -1,8 +1,8 @@
 import { handler } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { channelServiceWhitelistMiddleware } from "@/pdc/providers/middleware/index.ts";
-import { handlerConfig, ParamsInterface, ResultInterface } from "@/shared/certificate/find.contract.ts";
-import { alias } from "@/shared/certificate/find.schema.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "@/pdc/services/certificate/contracts/find.contract.ts";
+import { alias } from "@/pdc/services/certificate/contracts/find.schema.ts";
 import { mapCertForListHelper } from "../helpers/mapCertForListHelper.ts";
 import { CertificateRepositoryProviderInterfaceResolver } from "../interfaces/CertificateRepositoryProviderInterface.ts";
