+++ api/src/pdc/services/certificate/contracts/common/interfaces/CertificateBaseInterface.ts
