+++ api/src/pdc/services/certificate/contracts/common/interfaces/CertificateInterface.ts
