+++ api/src/pdc/services/certificate/contracts/common/interfaces/CertificateMetaInterface.ts
@@ -1,5 +1,5 @@
-import { PointInterface } from '../../../common/interfaces/PointInterface.ts';
-import { CarpoolInterface } from './CarpoolInterface.ts';
+import { PointInterface } from "../../../../../../shared/common/interfaces/PointInterface.ts";
+import { CarpoolInterface } from "./CarpoolInterface.ts";
 
 export interface MetaPersonDisplayInterface {
   datetime: Date;
