+++ api/src/pdc/services/certificate/helpers/mapFromCarpools.ts
@@ -1,12 +1,12 @@
 import {
   CarpoolInterface,
   CarpoolTypeEnum,
-} from "@/shared/certificate/common/interfaces/CarpoolInterface.ts";
-import { CertificateBaseInterface } from "@/shared/certificate/common/interfaces/CertificateBaseInterface.ts";
+} from "@/pdc/services/certificate/contracts/common/interfaces/CarpoolInterface.ts";
+import { CertificateBaseInterface } from "@/pdc/services/certificate/contracts/common/interfaces/CertificateBaseInterface.ts";
 import {
   CertificateMetaInterface,
   MetaPersonInterface,
-} from "@/shared/certificate/common/interfaces/CertificateMetaInterface.ts";
+} from "@/pdc/services/certificate/contracts/common/interfaces/CertificateMetaInterface.ts";
 import { PointInterface } from "@/shared/common/interfaces/PointInterface.ts";
 
 export interface ParamsInterface {
